Debugged app
A number of bugs have been removed with this
commit. 1) The event listener waiting for a
click on the old game board has been added to
the function that removes other event listeners
and thus avoids multiple games run in parallel.
2) The game loop has been made a bit dryer and
an if condition has been added to make the game
board setup at the outset only run for when the
game board actually has contents, that is when
and old game has been picked up instead of playing
a new one.