Add setup-new-game-page
This page will appear when the player chooses to
set up a new game (clicking on Play in the main
nav bar. The new page will give options to choose
the symbol, which turn to start, which opponent
to play, and, if a computer opponent is chosen,
which AI difficulty to play against.