Add option to pick symbol and turn order
The option to pick your symbol (X or O) has been
added in tandem with the option to pick which
order human players (X and O) want to start in.
The game-loop module has been refactored my
moving the game logic associated with human
players to the humanMove module and the logic
associated with computer players to the aiMove
module. Next step in the development process is
to code the ai algo.