AI and general debugging
Debugging the hard AI involved a number of steps,
including rebuilding the utility function and
ensuring that no variables inside the minmax
were driven by variables scoped outside the
function.

In addition, a bug was removed whereby a user
could have entered his new password, signed out,
and a new user, signing in on the same machine,
could have then exploited the password as it was
still in the change password form. This has been
removed by adding a function to the click event
callback of the sign-out button whereby all forms
on the SPA are deleted.